uvedit: floats were being implicitly promoted to doubles, adjust to use floats.

+++ source/blender/editors/uvedit/uvedit_ops.c
@@ -352,12 +352,12 @@ void uvedit_uv_deselect(Scene *scene, EditFace *efa, MTFace *tf, int i)
 void uv_center(float uv[][2], float cent[2], int quad)
 {
 	if(quad) {
-		cent[0] = (uv[0][0] + uv[1][0] + uv[2][0] + uv[3][0]) / 4.0;
-		cent[1] = (uv[0][1] + uv[1][1] + uv[2][1] + uv[3][1]) / 4.0;		
+		cent[0] = (uv[0][0] + uv[1][0] + uv[2][0] + uv[3][0]) / 4.0f;
+		cent[1] = (uv[0][1] + uv[1][1] + uv[2][1] + uv[3][1]) / 4.0f;
 	}
 	else {
-		cent[0] = (uv[0][0] + uv[1][0] + uv[2][0]) / 3.0;
-		cent[1] = (uv[0][1] + uv[1][1] + uv[2][1]) / 3.0;		
+		cent[0] = (uv[0][0] + uv[1][0] + uv[2][0]) / 3.0f;
+		cent[1] = (uv[0][1] + uv[1][1] + uv[2][1]) / 3.0f;
 	}
 }
 
@@ -436,8 +436,8 @@ static int uvedit_center(Scene *scene, Image *ima, Object *obedit, float *cent, 
 	}
 	
 	if(change) {
-		cent[0]= (min[0]+max[0])/2.0;
-		cent[1]= (min[1]+max[1])/2.0;
+		cent[0]= (min[0]+max[0])/2.0f;
+		cent[1]= (min[1]+max[1])/2.0f;
 		
 		BKE_mesh_end_editmesh(obedit->data, em);
 		return 1;
@@ -578,9 +578,9 @@ static void find_nearest_uv_vert(Scene *scene, Image *ima, EditMesh *em, float c
 
 			for(i=0; i<nverts; i++) {
 				if(penalty && uvedit_uv_selected(scene, efa, tf, i))
-					dist= fabs(co[0]-tf->uv[i][0])+penalty[0] + fabs(co[1]-tf->uv[i][1])+penalty[1];
+					dist= fabsf(co[0]-tf->uv[i][0])+penalty[0] + fabsf(co[1]-tf->uv[i][1]) + penalty[1];
 				else
-					dist= fabs(co[0]-tf->uv[i][0]) + fabs(co[1]-tf->uv[i][1]);
+					dist= fabsf(co[0]-tf->uv[i][0]) + fabsf(co[1]-tf->uv[i][1]);
 
 				if(dist<=mindist) {
 					if(dist==mindist)
@@ -1438,7 +1438,7 @@ static int sticky_select(float *limit, int hitv[4], int v, float *hituv[4], floa
 	for(i=0; i<4; i++) {
 		if(hitv[i] == v) {
 			if(sticky == SI_STICKY_LOC) {
-				if(fabs(hituv[i][0]-uv[0]) < limit[0] && fabs(hituv[i][1]-uv[1]) < limit[1])
+				if(fabsf(hituv[i][0]-uv[0]) < limit[0] && fabsf(hituv[i][1]-uv[1]) < limit[1])
 					return 1;
 			}
 			else if(sticky == SI_STICKY_VERTEX)
@@ -2280,7 +2280,7 @@ static void select_uv_inside_ellipse(Scene *scene, int select, EditFace *efa, MT
 	y= (uv[1] - offset[1])*ell[1];
 
 	r2 = x*x + y*y;
-	if(r2 < 1.0) {
+	if(r2 < 1.0f) {
 		if(select)	uvedit_uv_select(scene, efa, tface, select_index);
 		else uvedit_uv_deselect(scene, efa, tface, select_index);
 	}
@@ -3124,7 +3124,7 @@ static int set_tile_invoke(bContext *C, wmOperator *op, wmEvent *event)
 	y= event->y - ar->winrct.ymin;
 	UI_view2d_region_to_view(&ar->v2d, x, y, &fx, &fy);
 
-	if(fx>=0.0 && fy>=0.0 && fx<1.0 && fy<1.0) {
+	if(fx >= 0.0f && fy >= 0.0f && fx < 1.0f && fy < 1.0f) {
 		fx= fx*ima->xrep;
 		fy= fy*ima->yrep;
